Improve multi-targeting functionality for artifacts (#149)
When a project targets multiple target frameworks, we want the artifacts to be copied after the outer build and not for each inner build by default.  We also want the copy artifacts target to run after Pack if a pack is taking place.

Also, the NUPKG should be copied by default.