Support .NET 5 SDK by importing its props and targets
Previous versions of the .NET SDK did not require you to import them to get some of their build logic.  More and more of the logic is being moved out of what ships with MSBuild and into what ships in the SDK.  To that end, Traversal should import the Microsoft.NET.Sdk and just extend it.