Fix extensibility of Compile and CoreCompile (#225)
Compile was removed from BuildDependsOn to avoid having it run but this breaks some extensibility points that other SDKs might key off of.  This change brings back the dependency chain to run Compile and CoreCompile but has the bulk of their logic no-op.  This allows the extensibility to continue to work as expected.

Bumped version to 3.0 as this could be a breaking change.

Fixes #223