Clean up logging from NoTargets projects (#300)
* Clean up logging from NoTargets projects

A NoTargets project will still log (with high importance) something like:

> MyProject ->

Leaving a blank where the primary output would typically go. This looks weird. So this change removes that from being logged. It also disables (because it's the same property that controls it) copying of the primary output that we didn't build.

* Add unit test for change

* Add condition

Co-authored-by: Jeff Kluge <jeffkl@microsoft.com>