Traversal: Keep GetTargetPath in sync with Build
https://github.com/microsoft/MSBuildSdks/commit/1c02e77c5d96c50b803e348b6c3444b578d3dee0 I made the GetTargetPath target for traversal projects return nothing. This is actually wrong as msbuild projects are expected to return the same data in the Build and the GetTargetPath target.
If those two targets aren't in sync, a behavior difference is observable when a traversal project is P2Pd and the dependent project reads from the `CollectedBuildOutput` (which is returned by the `ResolveProjectReferences` target). If the `BuildProjectReferences=true` flag is passed in, the output in the ResolveP2P is currently empty vs. not empty during a normal build without the flag set.